refactor(util): adjust `ConcurrencyLimiter` util naming convention
