+++ packages/@sanity/util/.gitignore
@@ -13,7 +13,7 @@
 
 # Exports
 /client.js
-/ConcurrencyLimiter.js
+/concurrency-limiter.js
 /content.js
 /createSafeJsonParser.js
 /fs.js
