+++ packages/@sanity/util/test/ConcurrencyLimiter.test.ts
@@ -1,4 +1,4 @@
-import {ConcurrencyLimiter} from '../src/ConcurrencyLimiter'
+import {ConcurrencyLimiter} from '../src/concurrency-limiter'
 
 const tick = () => new Promise((resolve) => setTimeout(resolve, 0))
