+++ packages/sanity/src/core/validation/validateDocument.ts
@@ -9,7 +9,7 @@ import {
   type ValidationMarker,
 } from '@sanity/types'
 import {createClientConcurrencyLimiter} from '@sanity/util/client'
-import {ConcurrencyLimiter} from '@sanity/util/ConcurrencyLimiter'
+import {ConcurrencyLimiter} from '@sanity/util/concurrency-limiter'
 import {flatten, uniqBy} from 'lodash'
 import {concat, defer, from, lastValueFrom, merge, Observable, of} from 'rxjs'
 import {catchError, map, mergeAll, mergeMap, switchMap, toArray} from 'rxjs/operators'
